Afolabs is an organization who helps to survey building
Am an engineer specialised in the construction of building like homes, offices and shops. I reside in Ado Ekiti along Basiri road. I am also a surveyor whose job is also to look into the maintain and survey the construction of buildings for example the bricklaying work, the furniture work, tiling and e.t.c I am a team leader